Update repo folder layout
