+++ config/csb_default.ini
@@ -1,0 +1,37 @@
+[global]
+version = 3.25
+python_env = C:/Program Files/ArcGIS/Pro/bin/Python/envs/arcgispro-py3/python.exe
+cpu_prct = 0.97
+count0 = 2
+
+[folders]
+code = X:/CSB-Project/CSB_CODE/CSB-Run
+data = X:/CSB-Project/CSB_OUT
+split_rasters = X:/CSB-Project/CSB_INPUT/Split_v005
+log = X:/CSB-Project/CSB_OUT
+
+[prep_tile]
+gee_file = <data>/Creation/GEE/Year
+combine_gee = <data>/Creation/GEE/AreaTiles
+output_splits = <split_rasters>
+
+[create]
+create_folder = <data>/Creation/<runname>
+prepped_cdl = <creation_folder>/gee
+creation_data = <creation_folder>/run_data 
+creation_final =  <creation_folder>/out
+creation_logs = <creation_folder>/logs
+
+[prep]
+prep_folder = <data>/Prep/<runname>
+prep_data = <prep_folder>/run_data
+prep_final = <prep_folder>/out
+prep_logs = <prep_folder>/logs
+cnty_shp_file = X://IMAGES\NATIONAL/Outlines/US_outlines/US48_tl2020_ASD_CNTY_Albers.shp
+national_cdl_folder = X:\\IMAGES\\NATIONAL\\CDL\\US
+
+[distribute]
+distribute_folder = <data>/Distribution/<runname>
+distribute_data = <distribute_folder>/run_data
+distribute_final = <distribute_folder>/out
+distribute_logs = <distribute_folder>/logs
